Add reason for closing component
